# Speed up half tensors printing