--- a/torch/_tensor_str.py
+++ b/torch/_tensor_str.py
@@ -328,6 +328,15 @@ def _tensor_str(self, indent):
     if self.is_neg():
         self = self.resolve_neg()
 
+    # TODO: Remove me when `masked_select` is implemented for FP8
+    if self.dtype in [
+        torch.float8_e5m2,
+        torch.float8_e5m2fnuz,
+        torch.float8_e4m3fn,
+        torch.float8_e4m3fnuz,
+    ]:
+        self = self.float()
+
     if self.dtype.is_complex:
         # handle the conjugate bit
         self = self.resolve_conj()