@malfet 2024-12-03
> Also interesting about float8's - I wonder if they can be converted instead to fp16 or bf16 (especially if these kernels exist) without losing precision and make use of the improvements in this PR wrt printing. This can also save some memory as fp8 -> fp32 conversion probably is hungry...

Sure, I can convert float8 to float16, but also think it might worth an effort expanding a few more generic select ops to float8, see https://github.com/pytorch/pytorch/pull/141928